Refactor the builder to encapsulate throwing exceptions
Change the builder to implement the build method in order to encapsulate
the logic to create IllegalStateExceptions that occur from the
validator.  Create two abstract methods that the build method calls.
One constructs the validator so that the builder can create the
validator.  The second is the logic to construct the object the build
method returns.